Made front end responsive for all active layouts in both themes
