+++ lib/LoadCustom.dart
@@ -5,7 +5,7 @@ import 'package:flutter/material.dart';
 import 'DatabaseHelper.dart';
 import 'package:flutter_slidable/flutter_slidable.dart';
 import 'CustomLayout.dart';
-import 'globals.dart';
+import 'Theme.dart';
 
 List<String> tableList;
