+++ lib/WasdLayout.dart
@@ -2,7 +2,7 @@ import 'package:aio_pc_controller/globals.dart';
 import 'package:flutter/material.dart';
 import 'package:flutter/services.dart';
 import 'config.dart';
-import 'globals.dart';
+import 'Theme.dart';
 
 class WasdLayout extends StatefulWidget {
   @override
@@ -10,6 +10,8 @@ class WasdLayout extends StatefulWidget {
 }
 
 class _WasdLayoutState extends State<WasdLayout> {
+
+  int wdark = 0, adark = 0, shiftdark = 0, sdark = 0, ddark = 0, spacedark = 0;
   @override
   Widget build(BuildContext context) {
     SystemChrome.setPreferredOrientations([
@@ -27,13 +29,15 @@ class _WasdLayoutState extends State<WasdLayout> {
               top: 110.0,
               child: GestureDetector(
                 onPanStart: (_) {
+                  setState((){wdark = 1;});
                    _send('down&w');
                 },
                 onPanEnd: (_) {
+                  setState((){wdark = 0;});
                    _send('up&w');
                 },
                 child: Container(
-                  color: currentThemeColors.accentColor,
+                  color: currentThemeColors.buttonColor[wdark],
                   padding: const EdgeInsets.all(20.0),
                   child: Icon(
                     Icons.arrow_upward,
@@ -48,13 +52,15 @@ class _WasdLayoutState extends State<WasdLayout> {
               top: 220.0,
               child: GestureDetector(
                 onPanStart: (_) {
+                  setState((){sdark = 1;});
                    _send('down&s');
                 },
                 onPanEnd: (_) {
+                  setState((){sdark = 0;});
                    _send('up&s');
                 },
                 child: Container(
-                  color: currentThemeColors.accentColor,
+                  color: currentThemeColors.buttonColor[sdark],
                   padding: const EdgeInsets.all(20.0),
                   child: Icon(
                     Icons.arrow_downward,
@@ -69,13 +75,15 @@ class _WasdLayoutState extends State<WasdLayout> {
               top: 165.0,
               child: GestureDetector(
                 onPanStart: (_) {
+                  setState((){spacedark = 1;});
                    _send('down&space');
                 },
                 onPanEnd: (_) {
+                  setState((){spacedark = 0;});
                    _send('up&space');
                 },
                 child: Container(
-                  color: currentThemeColors.accentColor,
+                  color: currentThemeColors.buttonColor[spacedark],
                   padding: const EdgeInsets.all(20.0),
                   child: Icon(
                     Icons.space_bar,
@@ -90,13 +98,15 @@ class _WasdLayoutState extends State<WasdLayout> {
               top: 70.0,
               child: GestureDetector(
                 onPanStart: (_) {
+                  setState((){shiftdark = 1;});
                    _send('down&shift');
                 },
                 onPanEnd: (_) {
+                  setState((){shiftdark = 0;});
                    _send('up&shift');
                 },
                 child: Container(
-                  color: currentThemeColors.accentColor,
+                  color: currentThemeColors.buttonColor[shiftdark],
                   padding: const EdgeInsets.all(20.0),
                   child: Icon(
                     Icons.keyboard_capslock,
@@ -111,13 +121,15 @@ class _WasdLayoutState extends State<WasdLayout> {
               top: 165.0,
               child: GestureDetector(
                 onPanStart: (_) {
+                  setState((){ddark = 1;});
                    _send('down&d');
                 },
                 onPanEnd: (_) {
+                  setState((){ddark = 0;});
                    _send('up&d');
                 },
                 child: Container(
-                  color: currentThemeColors.accentColor,
+                  color: currentThemeColors.buttonColor[ddark],
                   padding: const EdgeInsets.all(20.0),
                   child: Icon(
                     Icons.arrow_forward,
@@ -132,13 +144,15 @@ class _WasdLayoutState extends State<WasdLayout> {
               top: 165.0,
               child: GestureDetector(
                 onPanStart: (_) {
+                  setState((){adark = 1;});
                    _send('down&a');
                 },
                 onPanEnd: (_) {
+                  setState((){adark = 0;});
                    _send('up&a');
                 },
                 child: Container(
-                  color: currentThemeColors.accentColor,
+                  color: currentThemeColors.buttonColor[adark],
                   padding: const EdgeInsets.all(20.0),
                   child: Icon(
                     Icons.arrow_back,
